fix wrong method calls.

+++ blkview/src/net/katsuster/blkview/LogRenderer.java
@@ -5,7 +5,7 @@ package net.katsuster.blkview;
  * @author katsuhiro
  */
 public interface LogRenderer {
-	public void setBlockCount(long n);
+	public void setBlockCount(int n);
 	public void setCapacity(long n);
 
 	public void addAccessLog(AccessLogRW log);
